// eshop-web/src/main/resources/client/src/app/oil/models/lang-info.model.ts
import { Constant } from 'src/app/core/conts/app.constant';

export class LangInfo {
    langCode: string = Constant.EMPTY_STRING;
    lang: string = Constant.EMPTY_STRING;
    countryLang: string = Constant.EMPTY_STRING;
    engineInfo: string = Constant.EMPTY_STRING;
    languageselectedvalue: string = Constant.EMPTY_STRING;
}